Container system finished. Added neat player picture.
Containers are all done, there is a bug with Layer container
when it's used as a parent for LibingBox, but that will be
fixed when time comes for Layer to be used, maybe I'll kick it
out of container tree totaly.

gamedesing/Container children.png - shows the current state of
the container system and what child implements what.

All the Entities have been reworked to fit into the system,
basically all of them turned out to be boxed entities.

Returned Player to the field and made it a cool custom icon.

Lines need to be reworked for easyer paint-mode switching.

Listeners require a rework and a split from InputListener
back to Keyboard and MouseListeners respectively.

BUG: Layer will not position relative to its x/y.
BIG: Game kinda "shakes"... I donno wut to do... O_o