Huge amount of work, read below.
1). MathExt, ColorExt are added.

MathExt implemetns rotation functionality and custom random(),

ColorExt got some needed color extensions like TRANSPARENT,
and random color generator.

2). Entity split into Entity -> EntityObject -> EntityLiving.

Entity has basic flags and a container. It's like a file.
Can be extended into inner game deamons like EntityDeathLord,
or DistanceFlicker for not updating the far-away entities.

EntityObject. This is still a questionable thing, but for now
it has everithing old entity got except movement.
Also holds the fillColor and the drawColor.

EntityLiving takes care of the movement, acceleration, friction,
active state. Can end up getting merged back into EntityObject.
Currently is the base class of all game's creatures.

3). Entity and Container recieved add() focused rework.

Currently when an entity is add()'ed it calls it's reset()
method. I don't like this.
=>
It will get scrapped in favor
of the constaructor calling the add() method by itself.
Due to this add() will be replaced with move() for general
needs, since an entity will be assigned a container from the
coment it's created.

4). Huge bugfixes.

New Rotor2 creature was used for an extensive bugtracing,
which resulted in fixing Rotor and Collider clumping bugs.
=>
The math is now ready to create CollidingBox and CollidingCircle.
Also the work may be started on making the CollidingOval.

And o'corse random bugfixing left and right.