ontainer rework. Now is Iterable. Entity retweaks.
Conainer reworked and now extends Iterable<Entity> for
the enchanced for loops used in iterating over it.

It now also has remove() and moveIn() methods, corresponding to
it's inner aditions and deletion arrays.

Additions and deletions are being synced with the container on the
update tick: deletions first additions later.

Entity merged back to the single class with EntityBasicAI now
extending it for friction / acceleration as well as some usefull
methods like hit(), isTouchingEdge() and so on.

Also, entity now (again) takes the conatiner in it's constructor,
but now it actually moveIn()'s itself to it's new home.

All basic creatures reworked for this purpose.

reset() scrapped from the face of earth. (for now)

BULLETS! and EXPLOSIONS!!!!! new creatures, great basic
functionality built in, grand opportunities for advancement
avaluble in the future.

Somegame is new.

More bugfixes.