Explosions now chain. Bullets will die to explosion.
Explosion is now Entity, NOT __EntityBasciAI__!
Explosion will kill all EntityBasicAI that it hits,
except Palyers. Bullet is an EntityBasicAI.

Entity's kill() is now boolean and returns false if entity
already dead. Inner dead flag is responsible for that.

EntityBasicIA now holds isCloseEntough() method. It takes in
coordinates and distance to check for.

Bullets can backfire XD
If you shoot a line of bullets and one is hit by expolsion
(such as the wall on-hit one) the explosions will chain
faster than bullet speed and wreck themself some Player ;)
(Do note that bullets and explosions do nothing to player.)